Added comments to the solution and renamed files
