+++ Hackerrank_problems/Merge_the_Tools/problem.md
